requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
alembic
python-dotenv
passlib[bcrypt]
python-jose
httpx 
pytest